+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `open_browser()`, which uses the `$BROWSER` environment variable before falling back to `open()`.
+- WSL-specific implementation. Previously, WSL used the same implementation as Linux. Now, the strategy is to use the
+system `xdg-open` if available, otherwise we try using the system's `wslview` command from
+[`wslu`](https://github.com/wslutilities/wslu).
+### Changed
+- On Linux (non-WSL), the system `xdg-open` is now used if present. Otherwise, the bundled version is used, as before.
+- The command name in the `OpenError::ExitStatus` variant is now returned as a `Cow<'static, str>` instead of a
+`&'static str`.
 ### Removed
 - `impl From<io::Error> for OpenError`.